requirements.txt
streamlit
pandas
biopython
altair
matplotlib
seaborn
